DataCollectorScript/requirements.txt
python-binance
psycopg2
pandas
